+++ app/src/play/java/chat/rocket/android/analytics/GoogleAnalyticsForFirebase.kt
@@ -36,13 +36,14 @@ class GoogleAnalyticsForFirebase @Inject constructor(val context: Context) :
 
     // WIDECHAT tracking BSSID
     override fun logMessageSent(event: SubscriptionTypeEvent, serverUrl: String) {
-        val bssid = SharedPreferenceHelper.getString(Constants.CURRENT_BSSID, "none")
-        val loc_permission = SharedPreferenceHelper.getString(Constants.LOCATION_PERMISSION, "none")
-        firebaseAnalytics.logEvent("message_sent", Bundle(3).apply {
+        var bssid = SharedPreferenceHelper.getString(Constants.CURRENT_BSSID, "none")
+        if (bssid == "none") {
+            bssid = SharedPreferenceHelper.getString(Constants.LOCATION_PERMISSION, "none")
+        }
+        firebaseAnalytics.logEvent("message_sent", Bundle(2).apply {
             putString("subscription_type", event.subscriptionTypeName)
 //            putString("server", serverUrl)
             putString("wifi_bssid", bssid)
-            putString("location_permission", loc_permission)
         })
     }
